refactor(terminal): unify input shortcut policy
Route direct and managed terminal key handling through a shared policy layer so clipboard shortcuts, window accelerators, and PTY input encoding stay aligned.

Fixes #187
